+++ app/util/smart-transactions/smart-publish-hook.ts
